+++ newrelic/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file documents changes to the `workbench-newrelic` library, including notes on how to upgrade to new versions.
 
+## 0.2
+
+### Added
+- Add `NewRelicMetrics` trait
+- Add `NewRelicMetricsInterpreter`
+- Add `FakeNewRelicMetricsInterpreter`
+
+### Changed
+- Changed constructor to `NewRelicMetrics.fromNewRelic(appName)`
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-newrelic" % "TRAVIS-REPLACE-ME"`
+
 ## 0.1
 
 ### Added